Add client-events.js which handles irc event > bridge events
Outline the algorithm used when there is no bot present.
