+++ lib/irclib/client.js
@@ -4,12 +4,10 @@ var q = require("q");
 var util = require("util");
 var EventEmitter = require("events").EventEmitter;
 var ident = require("./ident");
-var pool = require("./client-pool");
-var actions = require("../models/actions");
 var clientConnection = require("./client-connection");
+var clientEvents = require("./client-events");
 var names = require("./names");
 var IrcRoom = require("../models/rooms").IrcRoom;
-var IrcUser = require("../models/users").IrcUser;
 var log = require("../logging").get("irc-client");
 
 // The length of time to wait before trying to join the channel again
@@ -304,7 +302,6 @@ BridgedClient.prototype.getLastActionTs = function() {
 };
 BridgedClient.prototype._onConnectionCreated = function(connInst, nameInfo, callbacks) {
     var self = this;
-    var server = this.server;
 
     // listen for a connect event which is done when the TCP connection is
     // established and set ident info (this is different to the connect() callback
@@ -323,164 +320,7 @@ BridgedClient.prototype._onConnectionCreated = function(connInst, nameInfo, call
         self.emit("client-disconnected", self);
     };
 
-    var createUser = function(nick) {
-        return new IrcUser(
-            server, nick,
-            pool.getBridgedClientByNick(server, nick) !== undefined
-        );
-    };
-
-    // === Attach bot listeners ===
-    if (this.isBot) {
-        // make the bot listen for join/parts
-        connInst.addListener("part", function(chan, nick, reason, msg) {
-            callbacks.onPart(server, createUser(nick), chan, "part");
-        });
-        connInst.addListener("quit", function(nick, reason, chans, msg) {
-            chans = chans || [];
-            chans.forEach(function(chan) {
-                callbacks.onPart(server, createUser(nick), chan, "quit");
-            });
-        });
-        connInst.addListener("kick", function(chan, nick, by, reason, msg) {
-            callbacks.onPart(server, createUser(nick), chan, "kick");
-        });
-        connInst.addListener("join", function(chan, nick, msg) {
-            callbacks.onJoin(server, createUser(nick), chan, "join");
-        });
-        // bucket names and drain them once per second to avoid flooding
-        // the matrix side with registrations / joins
-        var namesBucket = [
-        //  { chan: <channel>, nick: <nick> }
-        ];
-        var processingBucket = false;
-        var popName = function() {
-            var name = namesBucket.pop(); // LIFO but who cares
-            if (!name) {
-                processingBucket = false;
-                return;
-            }
-            self.log.debug(
-                "Pop %s/%s from names bucket (%s remaining)",
-                name.nick, name.chan, namesBucket.length
-            );
-            return callbacks.onJoin(
-                server, createUser(name.nick), name.chan, "names"
-            );
-        };
-        var purgeNames = function() {
-            var promise = popName();
-            if (promise) {
-                promise.finally(function() {
-                    purgeNames();
-                });
-            }
-        };
-
-        connInst.addListener("names", function(chan, names, msg) {
-            if (names) {
-                var userlist = Object.keys(names);
-                userlist.forEach(function(nick) {
-                    namesBucket.push({
-                        chan: chan,
-                        nick: nick
-                    });
-                    // var opsLevel = names[nick]; // + @ or empty string
-                    // TODO do something with opsLevel
-                });
-                self.log.info(
-                    "NAMEs: Adding %s nicks from %s.", userlist.length, chan
-                );
-                self.log.debug("Names bucket has %s entries", namesBucket.length);
-                if (!processingBucket) {
-                    processingBucket = true;
-                    purgeNames();
-                }
-            }
-        });
-        // listen for mode changes
-        connInst.addListener("+mode", function(channel, by, mode, arg) {
-            callbacks.onMode(server, channel, by, mode, true, arg);
-        });
-        connInst.addListener("-mode", function(channel, by, mode, arg) {
-            callbacks.onMode(server, channel, by, mode, false, arg);
-        });
-        connInst.addListener("message", function(from, to, text) {
-            if (to.indexOf("#") !== 0) { return; }
-            callbacks.onMessage(
-                server, createUser(from), to,
-                actions.irc.createMessage(text)
-            );
-        });
-        connInst.addListener("ctcp-privmsg", function(from, to, text) {
-            if (to.indexOf("#") !== 0) { return; }
-            if (text.indexOf("ACTION ") === 0) {
-                callbacks.onMessage(
-                    server, createUser(from), to,
-                    actions.irc.createEmote(
-                        text.substring("ACTION ".length)
-                    )
-                );
-            }
-        });
-        connInst.addListener("notice", function(from, to, text) {
-            if (to.indexOf("#") !== 0) { return; }
-            if (from) { // ignore server notices
-                callbacks.onMessage(
-                    server, createUser(from), to,
-                    actions.irc.createNotice(text)
-                );
-            }
-        });
-        connInst.addListener("topic", function(channel, topic, nick) {
-            if (channel.indexOf("#") !== 0) { return; }
-
-            if (nick && nick.indexOf("@") !== -1) {
-                var match = nick.match(
-                    // https://github.com/martynsmith/node-irc/blob/master/lib/parse_message.js#L26
-                    /^([_a-zA-Z0-9\[\]\\`^{}|-]*)(!([^@]+)@(.*))?$/
-                );
-                if (match) {
-                    nick = match[1];
-                }
-            }
-            callbacks.onMessage(
-                server, createUser(nick), channel,
-                actions.irc.createTopic(topic)
-            );
-        });
-    }
-
-    // === Attach client listeners ===
-    if (!this.isBot) {
-        // listen for PMs for clients. If you listen for rooms, you'll get
-        // duplicates since the bot will also invoke the callback fn!
-        connInst.addListener("message", function(from, to, text) {
-            if (to.indexOf("#") === 0) { return; }
-            callbacks.onPrivateMessage(
-                server, createUser(from), createUser(to),
-                actions.irc.createMessage(text)
-            );
-        });
-        connInst.addListener("notice", function(from, to, text) {
-            if (!from || to.indexOf("#") === 0) { return; }
-            callbacks.onPrivateMessage(
-                server, createUser(from), createUser(to),
-                actions.irc.createNotice(text)
-            );
-        });
-        connInst.addListener("ctcp-privmsg", function(from, to, text) {
-            if (to.indexOf("#") === 0) { return; }
-            if (text.indexOf("ACTION ") === 0) {
-                callbacks.onPrivateMessage(
-                    server, createUser(from), createUser(to),
-                    actions.irc.createEmote(
-                        text.substring("ACTION ".length)
-                    )
-                );
-            }
-        });
-    }
+    clientEvents.addHooks(this, connInst, callbacks);
 };
 module.exports.BridgedClient = BridgedClient;
